src/util/MemorySize/MemorySize: Factory function bytes now works with every integral type.

+++ test/MemorySizeTest.cpp
@@ -389,22 +389,22 @@ TEST(MemorySize, ArithmeticOperatorsOverAndUnderFlow) {
   // Floating point multiplication.
   ASSERT_THROW(ad_utility::MemorySize::bytes(ad_utility::size_t_max) * 1.5,
                std::overflow_error);
-  ASSERT_NO_THROW(ad_utility::MemorySize::bytes(
-                      static_cast<float>(ad_utility::size_t_max) / 2.3) *
+  ASSERT_NO_THROW(ad_utility::MemorySize::bytes(static_cast<size_t>(
+                      static_cast<float>(ad_utility::size_t_max) / 2.3)) *
                   2.3);
   ad_utility::MemorySize memFloatingPointMultiplication{
       ad_utility::MemorySize::bytes(ad_utility::size_t_max)};
   ASSERT_THROW(memFloatingPointMultiplication *= 1.487, std::overflow_error);
   memFloatingPointMultiplication = ad_utility::MemorySize::bytes(
-      static_cast<float>(ad_utility::size_t_max) / 4.73);
+      static_cast<size_t>(static_cast<float>(ad_utility::size_t_max) / 4.73));
   ASSERT_NO_THROW(memFloatingPointMultiplication *= 4.73);
 
   // Floating point division. We are checking for overflow via divisor, that
   // results in a quotient bigger than the dividend. For example: 1/(1/2) = 2
   ASSERT_THROW(100_GB / (1. / static_cast<float>(ad_utility::size_t_max)),
                std::overflow_error);
-  ASSERT_NO_THROW(ad_utility::MemorySize::bytes(
-                      static_cast<float>(ad_utility::size_t_max) / 2.4) /
+  ASSERT_NO_THROW(ad_utility::MemorySize::bytes(static_cast<size_t>(
+                      static_cast<float>(ad_utility::size_t_max) / 2.4)) /
                   (1. / 2.4));
   ad_utility::MemorySize memFloatingPointDivision{12_MB};
   ASSERT_THROW(memFloatingPointDivision /=
